requirement.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose
python-multipart
cryptography
email-validator
PyJWT
motor pymongo
pytest
